+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## Unreleased
+
+### Fixed
+- Resource object type assertion caused a PHP error.
+
 ## [0.1.0] - 2017-09-02
 
 Initial commit of classes and tests brought in from `cloudcreativity/json-api` at `0.10.1`.